Окончил вынесение расширений в отдельный файл.
