+++ HM12/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
