+++ HM12/Resources/Assets.xcassets/Contents.json
