+++ HM12/Resources/Assets.xcassets/Авиарежим.imageset/Contents.json
