+++ HM12/Resources/Assets.xcassets/Авиарежим.imageset/airplane-mode.png
