+++ HM12/Resources/Assets.xcassets/Звуки, тактильные сигналы.imageset/Contents.json
