+++ HM12/Resources/Assets.xcassets/Основные.imageset/Contents.json
