+++ HM12/Resources/Assets.xcassets/Основные.imageset/settings.png
