+++ HM12/Resources/Assets.xcassets/Пункт управления.imageset/Contents.json
