+++ HM12/Resources/Assets.xcassets/Режим модема.imageset/modem.png
