+++ HM12/Resources/Assets.xcassets/Сотовая связь.imageset/Contents.json
