+++ HM12/Resources/Assets.xcassets/Фокусирование.imageset/Contents.json
