+++ HM12/Resources/Assets.xcassets/Фокусирование.imageset/moon.png
