+++ HM12/Resources/Assets.xcassets/Экран 'Домой'.imageset/homeScreen.png
