+++ HM12/Resources/Assets.xcassets/Экран и яркость.imageset/Contents.json
