+++ HM12/Resources/Assets.xcassets/Экранное время.imageset/sand-watch.png
