+++ HM12/Resources/Configs/Info.plist
