+++ HM12/Sources/Scenes/BluetothSettingsViewController/BluetothSettingsViewControllerDataSource.swift
@@ -1,58 +1,11 @@
 //
-//  BluetothSettingsViewController.swift
+//  BluetothSettingsViewControllerDataSource.swift
 //  HM12
 //
-//  Created by Александр Петрович on 15.12.2021.
+//  Created by Aliaksandr Piatrovich on 17.01.22.
 //
 
 import UIKit
-
-
-
-class BluetothSettingsViewController: UIViewController {
-    
-    static var bluetoothSettingsData = Section.getBluetoothOffData()
-    static var bluetoothOnSettingsData = Section.getBluetoothOnData()
-    
-    let bluetoothSwitch = UISwitch()
-    
-    private lazy var bluetoothTableView: UITableView = {
-        let tableView = UITableView(frame: .zero, style: .insetGrouped)
-        tableView.backgroundColor = .systemGray6
-        tableView.dataSource = self
-        tableView.delegate = self
-        return tableView
-    }()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationItem.largeTitleDisplayMode = .never
-        title = "Bluetooth"
-        view.backgroundColor = .systemGray6
-        viewHierarchy()
-        setupLayout()
-    }
-    
-    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-    
-    private func viewHierarchy() {
-        view.addSubview(bluetoothTableView)
-    }
-    
-    private func setupLayout() {
-        bluetoothTableView.translatesAutoresizingMaskIntoConstraints = false
-        bluetoothTableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
-        bluetoothTableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        bluetoothTableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        bluetoothTableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor).isActive = true
-    }
-}
 
 extension BluetothSettingsViewController: UITableViewDataSource {
     
@@ -118,19 +71,6 @@ extension BluetothSettingsViewController: UITableViewDataSource {
             BluetothSettingsViewController.bluetoothSettingsData = Section.getBluetoothOffData()
             settingModel[1].options[2].detailTextLabel = "Выкл."
             bluetoothSwitch.setOn(false, animated: true)
-            bluetoothTableView.reloadData()
-        }
-    }
-}
-
-extension BluetothSettingsViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        switch [indexPath.section, indexPath.row] {
-        case [0, 0]:
-            break
-        default:
-            BluetothSettingsViewController.bluetoothSettingsData[indexPath.section].options[indexPath.row].detailTextLabel = "Подключено"
-            BluetothSettingsViewController.bluetoothOnSettingsData[indexPath.section].options[indexPath.row].detailTextLabel = "Подключено"
             bluetoothTableView.reloadData()
         }
     }
