+++ HM12/Sources/Scenes/TableViewCell.swift
@@ -1,0 +1,14 @@
+//
+//  TableViewCell.swift
+//  HM12
+//
+//  Created by Aliaksandr Piatrovich on 17.01.22.
+//
+
+import UIKit
+
+class TableViewSell: UITableViewCell {
+    override var reuseIdentifier: String? {
+        return ViewController.idCell
+    }
+}
